% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlphaPart.ped}
\alias{AlphaPart.ped}
\title{Sample pedigree for partition.}
\format{
A data frame with 8 rows and 8 variables:
\describe{
\item{id}{individual's ID}
\item{father}{Father's ID}
\item{mother}{Mother's ID}
\item{generation}{Generation}
\item{country}{Country}
\item{sex}{Individual's sex}
\item{trait1}{Breeding value for trait 1}
\item{trait2}{Breeding value for trait 2}
}
}
\source{
Simulation.
}
\usage{
data(AlphaPart.ped)
}
\description{
A dataset containing pedigree information and traits
for eight individuals.
}
\examples{
data(AlphaPart.ped)
print(AlphaPart.ped)
}
\keyword{datasets}
