% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getScale}
\alias{getScale}
\title{Get scale information}
\description{
This is an internally called function
}
\seealso{
\code{\link[AlphaPart]{AlphaPart}}
}
\author{
Thiago de Paula Oliveira
}
\keyword{internal}
