% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMethods.R
\name{object.size}
\alias{object.size}
\alias{c.object_sizes}
\alias{as.object_sizes}
\alias{is.object_sizes}
\alias{format.object_sizes}
\alias{print.object_sizes}
\title{Report the Space Allocated for Objects}
\usage{
\method{print}{object_sizes}(
  x,
  quote = FALSE,
  humanReadable = getOption("humanReadable"),
  standard = "IEC",
  units,
  digits = 1,
  width = NULL,
  sep = " ",
  justify = c("right", "left"),
  ...
)

\method{format}{object_sizes}(
  x,
  humanReadable = getOption("humanReadable"),
  standard = "IEC",
  units,
  digits = 1,
  width = NULL,
  sep = " ",
  justify = c("right", "left"),
  ...
)

is.object_sizes(x)

as.object_sizes(x)

\method{c}{object_sizes}(..., recursive = FALSE)
}
\arguments{
\item{...}{additional arguments passed to methods}

\item{x}{object or objects whose size should be measured}

\item{quote}{logical; should strings be quoted?}

\item{humanReadable}{logical or NULL; should sizes be formatted in human-readable form?}

\item{standard}{character string; unit standard to use}

\item{units}{character; units to use in printing}

\item{digits}{integer; number of significant digits}

\item{width}{integer; width of printed output}

\item{sep}{character; separator used in printed output}

\item{justify}{character; justification for printing}

\item{recursive}{logical; whether to recurse when combining}
}
\description{
Provides an estimate of the memory that is being used to store \code{R} objects.
}
\seealso{
\code{\link[AlphaPart]{AlphaPart}}
}
