% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{savePlot}
\alias{savePlot}
\title{Save plot method for \code{AlphaPart}}
\usage{
savePlot(...)
}
\arguments{
\item{...}{Arguments passed to \code{type} specific methods, say
\code{width} and \code{height} for \code{type="pdf"} etc.}
}
\value{
Beside the side effect of saving plots to disk, filenames are printed on
screen during the process and at the end invisibly returned.
}
\description{
Save plot method for \code{AlphaPart}
}
