% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sEBV}
\alias{sEBV}
\title{Scale EBVs for objects of the class summaryAlphaPart.}
\description{
This is an internally called functions used to Scale
EBVs in respect to base population for objects of the class
\code{AlphaPart}.
}
\seealso{
\code{\link[AlphaPart]{AlphaPart}}
}
\author{
Thiago de Paula Oliveira
}
\keyword{internal}
