% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaPart.R
\docType{package}
\name{AlphaPart-package}
\alias{AlphaPart-package}
\title{AlphaPart: Partition of genetic values by paths}
\description{
AlphaPart partitions genetic values and their summaries to
quantify the sources of genetic change in multi-generational pedigrees.
The partitioning method is described in Garcia-Cortes et al. (2008)
\url{doi:10.1017/S175173110800205X}. This method enables a retrospective genetic
analysis of past changes in a population and is as such a counterpart to
prospective genetic methods describing genetic change .
The package includes the main function AlphaPart for
partitioning genetic values and auxiliary functions for manipulating data and
summarizing, visualizing, and saving results.
Genetic values can be breeding values, allele dosages, or similar quantities.

See the introductory vignette for instructions on using this package.
The vignette can be viewed using the following command:
\code{vignette("intro", package = "AlphaPart")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaGenes/AlphaPart}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

Other contributors:
\itemize{
  \item Jana Obsteter \email{jana.obsteter@kis.si} (\href{https://orcid.org/0000-0003-1511-3916}{ORCID}) [contributor]
  \item Thiago de Paula Oliveira \email{thiago.oliveira@ed.ac.uk} (\href{https://orcid.org/0000-0002-4555-2584}{ORCID}) [contributor]
  \item Ros Craddock \email{ros.craddock@ed.ac.uk} (\href{https://orcid.org/0009-0004-1578-1580}{ORCID}) [contributor]
}

}
\keyword{internal}
