% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AlphaPart.ped}
\alias{AlphaPart.ped}
\title{Sample pedigree for partition.}
\format{
A data frame with 6 rows and 8 variables:
\describe{
  \item{IId}{individual's ID}
  \item{FId}{Father's ID}
  \item{MId}{Mother's ID}
  \item{gen}{Generation}
  \item{country}{Country}
  \item{gender}{Individual's sex}
  \item{bv1}{Breeding value for trait 1}
  \item{bv2}{Breeding value for trait 1}
}
}
\source{
Simulation.
}
\usage{
AlphaPart.ped
}
\description{
A dataset containing pedigree information and breeding values
for six individuals.
}
\keyword{datasets}
