% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_22.R
\name{centerPop}
\alias{centerPop}
\title{Calculate parent average for base population.}
\description{
This is an internally called functions used to calculate
parent average for base population.
}
\seealso{
\code{\link[AlphaPart]{AlphaPart}}
}
\author{
Thiago de Paula Oliveira
}
