% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMethods.R
\name{object.size}
\alias{object.size}
\alias{c.object_sizes}
\alias{as.object_sizes}
\alias{is.object_sizes}
\alias{format.object_sizes}
\alias{print.object_sizes}
\title{Report the Space Allocated for Objects}
\usage{
\method{print}{object_sizes}(
  x,
  quote = FALSE,
  humanReadable = getOption("humanReadable"),
  standard = "IEC",
  units,
  digits = 1,
  width = NULL,
  sep = " ",
  justify = c("right", "left"),
  ...
)

\method{format}{object_sizes}(
  x,
  humanReadable = getOption("humanReadable"),
  standard = "IEC",
  units,
  digits = 1,
  width = NULL,
  sep = " ",
  justify = c("right", "left"),
  ...
)

is.object_sizes(x)

as.object_sizes(x)

\method{c}{object_sizes}(..., recursive = FALSE)
}
\description{
Provides an estimate of the memory that is being used to store \code{R} objects.
}
\seealso{
\code{\link[AlphaPart]{AlphaPart}}
}
\keyword{internal}
