% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plotSummaryAlphaPart.R
\name{print.plotSummaryAlphaPart}
\alias{print.plotSummaryAlphaPart}
\title{print.plotSummaryAlphaPart.R}
\usage{
\method{print}{plotSummaryAlphaPart}(x, ask = interactive(), ...)
}
\arguments{
\item{x}{plotSummaryAlphaPart, output object from
\code{\link[AlphaPart]{plot.summaryAlphaPart}} function}

\item{ask}{Logical, ask before printing another plot?}

\item{...}{Arguments passed to other functions (not used at the moment).}
}
\description{
#' Print method for objects of the class \code{plotSummaryAlphaPart} (result of
\code{plot(summary(AlphaPart(...)))}).
}
\details{
TODO: DETAILS
}
\examples{
## Load an example data
data("AlphaPart.ped")

## Partition additive genetic values
(res <- AlphaPart(x=AlphaPart.ped, colPath="loc", colAGV=c("agv1", "agv2")))

## Summarize population by generation (=trend)
(ret <- summary(res, by="gen"))

## Plot the partitions
p <- plot(ret, ylab=c("AGV for trait 1", "AGV for trait 2"), xlab="Generation")
print(p[[1]]$abs)
print(p[[1]]$rel)
print(p[[2]]$abs)
print(p[[2]])
print(p)
}
\seealso{
\code{\link[AlphaPart]{plot.summaryAlphaPart}}
}
