% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.csv.R
\name{write.csv}
\alias{write.csv}
\alias{write.csv.default}
\alias{write.csv.AlphaPart}
\alias{write.csv.summaryAlphaPart}
\title{Write Comma-Separated Values (CSV) file}
\usage{
write.csv(...)

\method{write.csv}{default}(...)

\method{write.csv}{AlphaPart}(x, file, traitsAsDir = FALSE, csv2 = TRUE, row.names = FALSE, ...)

\method{write.csv}{summaryAlphaPart}(x, file, traitsAsDir = FALSE, csv2 = TRUE, row.names = FALSE, ...)
}
\arguments{
\item{...}{other options passed to \code{\link[utils]{write.csv2}} or
\code{\link[utils]{write.csv}}.}

\item{x}{AlphaPart or summaryAlphaPart, objects returned from
\code{\link[AlphaPart]{AlphaPart}} or \code{\link[AlphaPart]{summary.AlphaPart}}}

\item{file}{character, file name with or without .csv extension,
that is, both \code{"file"} and \code{"file.csv"} are valid}

\item{traitsAsDir}{logical, should results be saved within trait folders;
named as \code{file.path(dirname(file), trait, basename(file))};
folders are created if they do not exist.}

\item{csv2}{logical, export using \code{\link[utils]{write.csv2}} or
\code{\link[utils]{write.csv}}.}

\item{row.names}{Logical, export row names as well?}
}
\value{
It contains:
\itemize{
\item \code{write.csv} - see \code{\link[utils]{write.csv}} for details.
\item \code{write.csv.AlphaPart} - for each trait (list component in \code{x})
a file is saved on disk with name \code{"AlphaPart_trait.csv"},
where the file will hold original data and genetic value partitions.
With \code{traitsAsDir=TRUE} files are saved as \code{"trait/file_trait.csv"}.
File names are printed on screen during the process of export and at the end
invisibly returned.
\item \code{`write.csv.summaryAlphaPart`} - for each trait (list component in \code{x})
a file partitions named \code{"file_trait.csv"} is saved on disk.
With \code{traitsAsDir=TRUE} files are saved as \code{"trait/file_trait_*.csv".}
File names are printed on screen during the process of export and at the end
invisibly returned.
}
}
\description{
Save summaries of partitioned genetic values to CSV files on disk for further
analyses of processing with other software or just for saving (backing up)
results.

Function \code{\link[utils]{write.csv}} from the \pkg{utils} package works
when the object is a \code{\link[base]{data.frame}} or a
\code{\link[base]{matrix}}. This is an attempt to make this function generic
so that one can define \code{write.csv} methods for other objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{write.csv(default)}: Default \code{write.csv} method.

\item \code{write.csv(AlphaPart)}: Save partitioned genetic values to CSV files on disk on disk
results.

\item \code{write.csv(summaryAlphaPart)}: Save summaries of partitioned genetic values to CSV files on disk

}}
\examples{
# Partition genetic values
data(AlphaPart.ped)
res <- AlphaPart(
  x = AlphaPart.ped,
  colPath = "country",
  colBV = c("trait1", "trait2")
)

# Write summary on the disk and collect saved file names
fileName <- file.path(tempdir(), "AlphaPart")
ret <- write.csv(x = res, file = fileName)
print(ret)
file.show(ret[1])

# Clean up
files <- dir(path = tempdir(), pattern = "AlphaPart*")
unlink(x = files)
}
\seealso{
\code{\link[utils]{write.csv}} help page on the default \code{write.csv}
and \code{write.csv2} methods in the \pkg{utils} package;
\code{\link[AlphaPart]{summary.AlphaPart}} and \code{\link[AlphaPart]{AlphaPart}}
help pages on \code{summaryAlphaPart} and \code{AlphaPart} classes.
}
